Move replicant binding to table into componenetDidMount callback
When I was working adding the role to the Player component, I was
experimenting with making it a part of the state which meant having
to create a relationship between that components state and the
playersRep replicant.  When I had that logic in the constructor,
I kept getting an error that I was setting the state of a component
that hadn't mounted, and it suggested using componentDidMount.

Resolves Ticket: WDL2