Change Codebase to Angular JS
This was a big change and a big commit. It would have been much
better to make this change it bite-sized chunks of commits, but
I didn't do that because I wanted this to feel like this represented
being at the same point in development as we were at before this
commit, just in a different framework.

Resolves Ticket: WDL6