Prompt broadcaster to designate whose cards were switched by troub.
In the course of implementing this user story, I discovered that
there was an issue with calling NodeCG.waitForReplicants too many
times.  It was warning of a memory leak and breaking the functionality
of my game phases.  To fix this, I saved the result of waitForReplicants
(a promise) as the constant 'safeToSet' and then called '.then' on that
constant wherever I wanted to do something after the replicants had
loaded.

Resolves Ticket: WDL17